codex: address PR review feedback (#20746)
