+++ codex-rs/tui/src/chatwidget/tests.rs
@@ -226,6 +226,7 @@ mod approval_requests;
 mod composer_submission;
 mod exec_flow;
 mod goal_menu;
+mod goal_validation;
 mod guardian;
 mod helpers;
 mod history_replay;
